// agency-services-transunion/lib/interfaces/merge-report.interface.ts
export interface IBankingRecord {
  BankingType?: ICodeRef;
  AccountDesignator?: ICodeRef;
  IndustryCode?: ICodeRef;
  Status?: ICodeRef;
  Remark?: IRemark | IRemark[];
  Source?: ISource;
  dateOpened?: string;
  dateClosed?: string;
  bureau?: string;
  dateVerified?: string;
  subscriberCode?: string;
  bankName?: string;
  balance?: number | string;
  accountNumber?: string;
}